Refactored chat history into new `_append_pr`
Also fixed user,assistant role loop in cases where a prompt is not
provided (e.g. `chat()`) but the previous message was an assistant. In
this case a user prompt "..." is added to allow the assistant to
continue.

This can be changed with `cont_pr` (and set to None to error out in
this case).